pyproject.toml
[project]
name = "experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "eclipse-zenoh>=0.11.0",
    "numpy>=1.20.0",
]
